+++ R/stat-qq.r
@@ -9,28 +9,38 @@
 #' @inheritParams layer
 #' @inheritParams geom_point
 #' @section Computed variables:
+#' Variables computed by \code{stat_qq}:
 #' \describe{
 #'   \item{sample}{sample quantiles}
 #'   \item{theoretical}{theoretical quantiles}
+#' }
+#' Variables computed by \code{stat_qq_line}:
+#' \describe{
+#'   \item{x}{x-coordinates of the endpoints of the line segment connecting the
+#'            points at the chosen quantiles of the theoretical and the sample
+#'            distributions}
+#'   \item{y}{y-coordinates of the endpoints}
 #' }
 #' @export
 #' @examples
 #' \donttest{
 #' df <- data.frame(y = rt(200, df = 5))
 #' p <- ggplot(df, aes(sample = y))
-#' p + stat_qq()
-#' p + geom_point(stat = "qq")
+#' p + stat_qq() + stat_qq_line()
 #'
 #' # Use fitdistr from MASS to estimate distribution params
 #' params <- as.list(MASS::fitdistr(df$y, "t")$estimate)
 #' ggplot(df, aes(sample = y)) +
-#'   stat_qq(distribution = qt, dparams = params["df"])
+#'   stat_qq(distribution = qt, dparams = params["df"]) +
+#'   stat_qq_line(distribution = qt, dparams = params["df"])
 #'
 #' # Using to explore the distribution of a variable
-#' ggplot(mtcars) +
-#'   stat_qq(aes(sample = mpg))
-#' ggplot(mtcars) +
-#'   stat_qq(aes(sample = mpg, colour = factor(cyl)))
+#' ggplot(mtcars, aes(sample = mpg)) +
+#'   stat_qq() +
+#'   stat_qq_line()
+#' ggplot(mtcars, aes(sample = mpg, colour = factor(cyl))) +
+#'   stat_qq() +
+#'   stat_qq_line()
 #' }
 geom_qq <- function(mapping = NULL, data = NULL,
                     geom = "point", position = "identity",